feat(team): add idempotent actor mailbox send
