+++ crates/agenthub-team-actor/src/lib.rs
@@ -5,8 +5,8 @@ mod transport;
 
 pub use mailbox::{
     AckActorMessageCommand, AckActorMessageResult, ActorMailbox, ActorMailboxError,
-    ActorMailboxStore, ListActorInboxQuery, PendingRemoteRelayRecord, RelayRemotePendingCommand,
-    RelayRemotePendingResult, SendActorMessageCommand,
+    ActorMailboxStore, CreatePendingMessageResult, ListActorInboxQuery, PendingRemoteRelayRecord,
+    RelayRemotePendingCommand, RelayRemotePendingResult, SendActorMessageCommand,
 };
 pub use message::{ActorMessageRecord, ActorMessageStatus, ActorMessageTransport};
 pub use relay::{ActorMessageRelay, ActorRelayError};
